Failed subtraction #208
- seems to be the same as #207, added
  test case
