+++ libs/calc-arithmetic/src/lib/positional/complement-extension.ts
@@ -1,4 +1,4 @@
-import { AdditionOperand, AdditionPositionResult, Digit } from '../models';
+import { AdditionOperand, AdditionPositionResult, Digit, PositionResult } from '../models';
 import { objArrayEqual } from '@calc/utils';
 import { BaseDigits } from './base-digits';
 
@@ -92,4 +92,44 @@ export function extendComplement<T extends Digit>(digits: T[], numPositions: num
     extensions[0].isComplementExtension = true;
 
     return [...extensions, ...rest];
+}
+
+export function mergeComplementExtension<T extends Digit>(resultDigits: T[], positionResults: PositionResult<T>[]): AdditionOperand[] {
+    const [, extensionDigit, ...rest] = resultDigits;
+    const firstDifferentIndex = findFirstNonRepeatingDigitIndex(resultDigits);
+
+    let startPositionIndex = firstDifferentIndex === -1
+        ? rest.length - 1
+        : firstDifferentIndex;
+
+    const positionIndexBeforeStart = startPositionIndex - 1;
+
+    const shouldStartFromPreviousPosition = startPositionIndex >= 1
+        && prevPositionGeneratedFromInitialDigits(startPositionIndex, rest, positionResults);
+
+    if (shouldStartFromPreviousPosition) startPositionIndex = positionIndexBeforeStart;
+
+    const startPosition = rest[startPositionIndex].position;
+    const mergedExtension = getComplementExtension(extensionDigit, startPosition + 1);
+    const nonExtensionDigits = rest.slice(startPositionIndex);
+
+    return [mergedExtension, ...nonExtensionDigits];
+}
+
+function prevPositionGeneratedFromInitialDigits<T extends Digit>(index: number, digits: AdditionOperand[], positionResults: PositionResult<T>[]): boolean {
+    const prevPosition = digits[index - 1].position;
+    const prevPositionResult = positionResults.find((res) => res.valueAtPosition.position === prevPosition);
+
+    return positionGeneratedFromInitialDigits(prevPositionResult);
+}
+
+function positionGeneratedFromInitialDigits<T extends Digit>(positionResult: PositionResult<T>): boolean {
+    return positionResult.operands.every((op) => !op.isComplementExtension);
+}
+
+export function findFirstNonRepeatingDigitIndex<T extends Digit>(resultDigits: T[]): number {
+    const [, extensionDigit, ...rest] = resultDigits;
+    return rest.findIndex((digit) => {
+        return digit.valueInDecimal != extensionDigit.valueInDecimal;
+    });
 }
